requirements.txt
fastapi
uvicorn
python-multipart
agno
google-generativeai
pillow
python-dotenv
streamlit
requests
